refactor: container 제거
